修改熱處理價格表（刪掉種類） 新增金屬價格表

+++ src/main/java/tw/idv/erp/order/order_detail/entity/OrderDetail.java
@@ -29,13 +29,13 @@ public class OrderDetail extends Core {
     private Integer categoryId;
 
     @Column(name = "length")
-    private Integer length;
+    private BigDecimal length;
 
     @Column(name = "width")
-    private Integer width;
+    private BigDecimal width;
 
     @Column(name = "thickness")
-    private Integer thickness;
+    private BigDecimal thickness;
 
     @Size(max = 50)
     @Column(name = "productMaterial", length = 50)
